Change DBRepair.sh to support Apple Mac.
This is the primary commit for PlexDBRepair that adds Mac support
to `DBRepair.sh`

Add all the PMS paths unique to Mac in `HostConfig()`.
Include code to create a `plexmediaserver.pid` file.
That concept needs more attention to guard against PMS running.

Deal with MacOS quirks in shell utils and builtins like
- sed -i
- echo -n
- pidof
- awk
- colons & forward slashes in filenames