[Windows] Check PMS state in more places
While it probably won't address #184 since the batch script works without issue,
but I was unable to replicate this on my system, so this is at least an
intermediate step. Also do some refactoring around the export process, and don't
remove temp files in scripted mode if the last operation failed.